Componentize autoscaler simulator code
Previously the autoscaler simulator lived entirely in a single file.
This commit decomposes the simulator into distinct components in order
to allow further enhancements of the simulator's flexibility and
features.

- Extract cmd/simulator
- Extract scenarios
  * Introduce SimulationConfiguration interface
  * Introduce ReplicaModel interface
  * Extract CombinedScenario
