+++ function-controller/Makefile
@@ -45,6 +45,6 @@ kubectl-apply:
 	kubectl apply -n riff-system -f config/deployment.yaml
 
 simulation:
-	go build pkg/controller/autoscaler/simulator/simulator.go
+	go build cmd/simulator/simulator.go
 	./simulator
 	gnuplot -c pkg/controller/autoscaler/simulator/simulator.plt > simulation-results.png
