+++ function-controller/pkg/controller/autoscaler/simulator/scenarios/scenarios_suite_test.go
@@ -1,4 +1,4 @@
-package main
+package scenarios
 
 import (
 	"testing"
